+++ CHANGES.rst
@@ -7,8 +7,8 @@ The change log for the previous version, Zope 2.13, is at
 https://zope.readthedocs.io/en/2.13/CHANGES.html
 
 
-4.4.5 (unreleased)
-------------------
+4.5 (unreleased)
+----------------
 
 - Make "Unicode Conflict Resolution" available for templates
   rendered with ``chameleon``